committing Leslie changes
Updated text, and made all Shiny Apps into just a link to the App, as to not waste hours, we are at 71 hours, so the two days I was out in the field about 17-20 hours were used. Hopefully this reduces it, but I will also do a timeout on the Shiny.io dashboard.